Create README.md
My name is Dre and I have decided to learn code. If you had asked me 5 years ago would i be interested in learning how to code the answer would definitely be no. But here I am learning how to code to eventually reach my long term goal of being a smart contract Dev. Long story short what brought me here is I was deployed to Greece and met a lovely  Russian woman  who we will called "mother Russia"  We went on a dinner date and unfortunately I was 10 euros off to being to pay for the bill. She told me not to worry about it and in response I promised to pay her back . However when trying to pay her back it was almost an impossible process due to the fact that Russia is banned from using the swift payment system.  For those of you who do not know what Swift is, Its pretty much a world wide transaction protocol that handles all financial transfers between western countries and their allies. So if a country or organization does an action that the west(US/UK) deems wrong then that country or organization is sanctioned, this also means that the people of that government are also sanctioned so they cant use normal measures to perform transactions that we can such as cash app, paypal, and so forth. I strongly believe that the people of that country should not be punished for the actions of an government or organizations. Hence here I am.